Make all demos end in suffix "Demo"
