+++ richtextfx-demos/src/main/java/org/fxmisc/richtext/demo/JavaKeywordsAsyncDemo.java
@@ -3,10 +3,14 @@ package org.fxmisc.richtext.demo;
 import java.time.Duration;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.Optional;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import javafx.application.Application;
+import javafx.concurrent.Task;
 import javafx.scene.Scene;
 import javafx.scene.layout.StackPane;
 import javafx.stage.Stage;
@@ -18,7 +22,7 @@ import org.fxmisc.richtext.model.StyleSpans;
 import org.fxmisc.richtext.model.StyleSpansBuilder;
 import org.reactfx.Subscription;
 
-public class JavaKeywords extends Application {
+public class JavaKeywordsAsyncDemo extends Application {
 
     private static final String[] KEYWORDS = new String[] {
             "abstract", "assert", "boolean", "break", "byte",
@@ -52,26 +56,26 @@ public class JavaKeywords extends Application {
     );
 
     private static final String sampleCode = String.join("\n", new String[] {
-        "package com.example;",
-        "",
-        "import java.util.*;",
-        "",
-        "public class Foo extends Bar implements Baz {",
-        "",
-        "    /*",
-        "     * multi-line comment",
-        "     */",
-        "    public static void main(String[] args) {",
-        "        // single-line comment",
-        "        for(String arg: args) {",
-        "            if(arg.length() != 0)",
-        "                System.out.println(arg);",
-        "            else",
-        "                System.err.println(\"Warning: empty string as argument\");",
-        "        }",
-        "    }",
-        "",
-        "}"
+            "package com.example;",
+            "",
+            "import java.util.*;",
+            "",
+            "public class Foo extends Bar implements Baz {",
+            "",
+            "    /*",
+            "     * multi-line comment",
+            "     */",
+            "    public static void main(String[] args) {",
+            "        // single-line comment",
+            "        for(String arg: args) {",
+            "            if(arg.length() != 0)",
+            "                System.out.println(arg);",
+            "            else",
+            "                System.err.println(\"Warning: empty string as argument\");",
+            "        }",
+            "    }",
+            "",
+            "}"
     });
 
 
@@ -79,37 +83,58 @@ public class JavaKeywords extends Application {
         launch(args);
     }
 
+    private CodeArea codeArea;
+    private ExecutorService executor;
+
     @Override
     public void start(Stage primaryStage) {
-        CodeArea codeArea = new CodeArea();
-
-        // add line numbers to the left of area
+        executor = Executors.newSingleThreadExecutor();
+        codeArea = new CodeArea();
         codeArea.setParagraphGraphicFactory(LineNumberFactory.get(codeArea));
-
-        // recompute the syntax highlighting 500 ms after user stops editing area
-        Subscription cleanupWhenNoLongerNeedIt = codeArea
-
-                // plain changes = ignore style changes that are emitted when syntax highlighting is reapplied
-                // multi plain changes = save computation by not rerunning the code multiple times
-                //   when making multiple changes (e.g. renaming a method at multiple parts in file)
-                .multiPlainChanges()
-
-                // do not emit an event until 500 ms have passed since the last emission of previous stream
+        Subscription cleanupWhenDone = codeArea.multiPlainChanges()
                 .successionEnds(Duration.ofMillis(500))
+                .supplyTask(this::computeHighlightingAsync)
+                .awaitLatest(codeArea.multiPlainChanges())
+                .filterMap(t -> {
+                    if(t.isSuccess()) {
+                        return Optional.of(t.get());
+                    } else {
+                        t.getFailure().printStackTrace();
+                        return Optional.empty();
+                    }
+                })
+                .subscribe(this::applyHighlighting);
 
-                // run the following code block when previous stream emits an event
-                .subscribe(ignore -> codeArea.setStyleSpans(0, computeHighlighting(codeArea.getText())));
-
-        // when no longer need syntax highlighting and wish to clean up memory leaks
-        // run: `cleanupWhenNoLongerNeedIt.unsubscribe();`
+        // call when no longer need it: `cleanupWhenFinished.unsubscribe();`
 
         codeArea.replaceText(0, 0, sampleCode);
 
         Scene scene = new Scene(new StackPane(new VirtualizedScrollPane<>(codeArea)), 600, 400);
-        scene.getStylesheets().add(JavaKeywordsAsync.class.getResource("java-keywords.css").toExternalForm());
+        scene.getStylesheets().add(JavaKeywordsAsyncDemo.class.getResource("java-keywords.css").toExternalForm());
         primaryStage.setScene(scene);
-        primaryStage.setTitle("Java Keywords Demo");
+        primaryStage.setTitle("Java Keywords Async Demo");
         primaryStage.show();
+    }
+
+    @Override
+    public void stop() {
+        executor.shutdown();
+    }
+
+    private Task<StyleSpans<Collection<String>>> computeHighlightingAsync() {
+        String text = codeArea.getText();
+        Task<StyleSpans<Collection<String>>> task = new Task<StyleSpans<Collection<String>>>() {
+            @Override
+            protected StyleSpans<Collection<String>> call() throws Exception {
+                return computeHighlighting(text);
+            }
+        };
+        executor.execute(task);
+        return task;
+    }
+
+    private void applyHighlighting(StyleSpans<Collection<String>> highlighting) {
+        codeArea.setStyleSpans(0, highlighting);
     }
 
     private static StyleSpans<Collection<String>> computeHighlighting(String text) {
